m2_item_manufacturer структура откорректирована

+++ var/dump/m2_item_manufacturer.strc.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `m2_item_manufacturer` (
-  `id` smallint(5) unsigned NOT NULL AUTO_INCREMENT,
-  `item_id` smallint(5) unsigned NOT NULL,
-  `order` smallint(5) unsigned NOT NULL,
+  `id` int(11) unsigned NOT NULL AUTO_INCREMENT,
+  `item_id` int(11) unsigned NOT NULL,
+  `order` int(11) unsigned NOT NULL,
   `manufacturer_id` int(11) unsigned NOT NULL,
   PRIMARY KEY (`id`),
   KEY `ssgn_project_id` (`item_id`),
